+++ Whistle/UI/Screens/Profile/MemberProfile/SEMemberProfileView.swift
@@ -79,7 +79,7 @@ struct SEMemberProfileView: View {
             ], spacing: 20) {
               ForEach(Array(apiViewModel.memberFeed.enumerated()), id: \.element) { index, content in
                 NavigationLink {
-                  MemberFeedKitView(index: index, userId: apiViewModel.memberFeed[index].userId ?? 0)
+                  MemberFeedView(index: index, userId: apiViewModel.memberFeed[index].userId ?? 0)
                     .id(UUID())
                 } label: {
                   videoThumbnailView(
